Use enum bulk_t for variable bulk_mode.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@2100 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/main.c
@@ -112,7 +112,7 @@ int classify(int argc, char **argv, FILE *out)
 	    break;
 	}
 	default:		/* command line mode */
-	    if (bulk_mode < argc && !error) {
+	    if ((int)bulk_mode < argc && !error) {
 		filename = argv[bulk_mode++];
 	    }
 	    else {
